requirements.txt
uvicorn
fastapi
sqlalchemy
python-dotenv
psycopg2
alembic
python-jose
passlib
python-multipart